extract list section header
